refactor: improved package naming and structure
