+++ src/main/java/com/synopsys/integration/detect/tool/detector/factory/DetectorFactory.java
@@ -15,21 +15,21 @@ import com.synopsys.integration.detect.tool.cache.InstalledToolLocator;
 import com.synopsys.integration.detect.tool.cache.InstalledToolManager;
 import com.synopsys.integration.detect.tool.detector.executable.DetectExecutableResolver;
 import com.synopsys.integration.detect.tool.detector.executable.DetectExecutableRunner;
-import com.synopsys.integration.detect.tool.detector.inspectors.ArtifactoryDockerInspectorResolver;
-import com.synopsys.integration.detect.tool.detector.inspectors.ArtifactoryGradleInspectorResolver;
-import com.synopsys.integration.detect.tool.detector.inspectors.ArtifactoryZipInstaller;
-import com.synopsys.integration.detect.tool.detector.inspectors.DockerInspectorInstaller;
-import com.synopsys.integration.detect.tool.detector.inspectors.LocalPipInspectorResolver;
-import com.synopsys.integration.detect.tool.detector.inspectors.nuget.AirgapNugetInspectorResolver;
-import com.synopsys.integration.detect.tool.detector.inspectors.nuget.ArtifactoryNugetInspectorInstaller;
-import com.synopsys.integration.detect.tool.detector.inspectors.nuget.NugetInspectorExecutableLocator;
-import com.synopsys.integration.detect.tool.detector.inspectors.nuget.OnlineNugetInspectorResolver;
-import com.synopsys.integration.detect.tool.detector.inspectors.projectinspector.AirgapProjectInspectorResolver;
-import com.synopsys.integration.detect.tool.detector.inspectors.projectinspector.ArtifactoryProjectInspectorInstaller;
-import com.synopsys.integration.detect.tool.detector.inspectors.projectinspector.LocalProjectInspectorInstaller;
-import com.synopsys.integration.detect.tool.detector.inspectors.projectinspector.OnlineProjectInspectorResolver;
-import com.synopsys.integration.detect.tool.detector.inspectors.projectinspector.ProjectInspectorExecutableLocator;
-import com.synopsys.integration.detect.tool.detector.inspectors.projectinspector.ProjectInspectorInstaller;
+import com.synopsys.integration.detect.tool.detector.inspector.ArtifactoryDockerInspectorResolver;
+import com.synopsys.integration.detect.tool.detector.inspector.ArtifactoryGradleInspectorResolver;
+import com.synopsys.integration.detect.tool.detector.inspector.ArtifactoryZipInstaller;
+import com.synopsys.integration.detect.tool.detector.inspector.DockerInspectorInstaller;
+import com.synopsys.integration.detect.tool.detector.inspector.LocalPipInspectorResolver;
+import com.synopsys.integration.detect.tool.detector.inspector.nuget.AirgapNugetInspectorResolver;
+import com.synopsys.integration.detect.tool.detector.inspector.nuget.ArtifactoryNugetInspectorInstaller;
+import com.synopsys.integration.detect.tool.detector.inspector.nuget.NugetInspectorExecutableLocator;
+import com.synopsys.integration.detect.tool.detector.inspector.nuget.OnlineNugetInspectorResolver;
+import com.synopsys.integration.detect.tool.detector.inspector.projectinspector.AirgapProjectInspectorResolver;
+import com.synopsys.integration.detect.tool.detector.inspector.projectinspector.OnlineProjectInspectorResolver;
+import com.synopsys.integration.detect.tool.detector.inspector.projectinspector.ProjectInspectorExecutableLocator;
+import com.synopsys.integration.detect.tool.detector.inspector.projectinspector.installer.ArtifactoryProjectInspectorInstaller;
+import com.synopsys.integration.detect.tool.detector.inspector.projectinspector.installer.LocalProjectInspectorInstaller;
+import com.synopsys.integration.detect.tool.detector.inspector.projectinspector.installer.ProjectInspectorInstaller;
 import com.synopsys.integration.detect.workflow.ArtifactResolver;
 import com.synopsys.integration.detect.workflow.airgap.AirGapInspectorPaths;
 import com.synopsys.integration.detect.workflow.file.DirectoryManager;
