+++ src/main/java/com/synopsys/integration/detect/tool/detector/inspector/LocalPipInspectorResolver.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.detect.tool.detector.inspectors;
+package com.synopsys.integration.detect.tool.detector.inspector;
 
 import java.io.File;
 import java.io.IOException;
