+++ src/main/java/com/synopsys/integration/detect/tool/detector/inspector/projectinspector/installer/ArtifactoryProjectInspectorInstaller.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.detect.tool.detector.inspectors.nuget;
+package com.synopsys.integration.detect.tool.detector.inspector.projectinspector.installer;
 
 import java.io.File;
 import java.io.IOException;
@@ -7,41 +7,43 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import com.synopsys.integration.detect.configuration.DetectInfo;
-import com.synopsys.integration.detect.tool.detector.inspectors.ArtifactoryZipInstaller;
+import com.synopsys.integration.detect.tool.detector.inspector.ArtifactoryZipInstaller;
+import com.synopsys.integration.detect.tool.detector.inspector.projectinspector.ProjectInspectorExecutableLocator;
 import com.synopsys.integration.detect.workflow.ArtifactoryConstants;
 import com.synopsys.integration.detectable.detectable.exception.DetectableException;
 import com.synopsys.integration.exception.IntegrationException;
 import com.synopsys.integration.util.OperatingSystemType;
 
-public class ArtifactoryNugetInspectorInstaller {
+public class ArtifactoryProjectInspectorInstaller implements ProjectInspectorInstaller {
     private final DetectInfo detectInfo;
     private final ArtifactoryZipInstaller artifactoryZipInstaller;
-    private final NugetInspectorExecutableLocator executableLocator;
+    private final ProjectInspectorExecutableLocator projectInspectorExecutableLocator;
 
-    public ArtifactoryNugetInspectorInstaller(
+    public ArtifactoryProjectInspectorInstaller(
         DetectInfo detectInfo,
-        ArtifactoryZipInstaller artifactoryZipInstaller, NugetInspectorExecutableLocator executableLocator
+        ArtifactoryZipInstaller artifactoryZipInstaller, ProjectInspectorExecutableLocator projectInspectorExecutableLocator
     ) {
         this.detectInfo = detectInfo;
         this.artifactoryZipInstaller = artifactoryZipInstaller;
-        this.executableLocator = executableLocator;
+        this.projectInspectorExecutableLocator = projectInspectorExecutableLocator;
     }
 
+    @Override
     @Nullable
     public File install(File directory) throws DetectableException {
         if (detectInfo.getCurrentOs() == OperatingSystemType.WINDOWS) {
-            return install(directory, ArtifactoryConstants.NUGET_INSPECTOR_WINDOWS_PROPERTY);
+            return install(directory, ArtifactoryConstants.PROJECT_INSPECTOR_WINDOWS_PROPERTY);
         } else if (detectInfo.getCurrentOs() == OperatingSystemType.MAC) {
-            return install(directory, ArtifactoryConstants.NUGET_INSPECTOR_MAC_PROPERTY);
+            return install(directory, ArtifactoryConstants.PROJECT_INSPECTOR_MAC_PROPERTY);
         } else {
-            return install(directory, ArtifactoryConstants.NUGET_INSPECTOR_LINUX_PROPERTY);
+            return install(directory, ArtifactoryConstants.PROJECT_INSPECTOR_LINUX_PROPERTY);
         }
     }
 
     @Nullable
     public File install(File installDirectory, String property) throws DetectableException {
         File extractedZip = downloadZip(property, installDirectory);
-        return executableLocator.findExecutable(extractedZip);
+        return projectInspectorExecutableLocator.findExecutable(extractedZip);
     }
 
     @NotNull //Returns location of extracted zip or throws
@@ -51,7 +53,7 @@ public class ArtifactoryNugetInspectorInstaller {
                 installDirectory,
                 ".zip",
                 ArtifactoryConstants.ARTIFACTORY_URL,
-                ArtifactoryConstants.NUGET_INSPECTOR_PROPERTY_REPO,
+                ArtifactoryConstants.PROJECT_INSPECTOR_PROPERTY_REPO,
                 property
             );
         } catch (IntegrationException | IOException e) {
